organized components and deleted CafeListing
